+++ packages/alfa-style/src/style.ts
@@ -189,12 +189,12 @@ export class Style implements Serializable {
 
             // https://drafts.csswg.org/css-cascade/#inherit
             case "inherit":
-              return this.inherited(name, source);
+              return this.inherited(name);
 
             // https://drafts.csswg.org/css-cascade/#inherit-initial
             case "unset":
               return inherits
-                ? this.inherited(name, source)
+                ? this.inherited(name)
                 : this.initial(name, source);
           }
         }
@@ -238,17 +238,8 @@ export class Style implements Serializable {
     return Value.of(Property.get(name).initial as Style.Computed<N>, source);
   }
 
-  public inherited<N extends Name>(
-    name: N,
-    source: Option<Declaration> = None
-  ): Value<Style.Inherited<N>> {
-    const inherited = this.parent.computed(name);
-
-    if (source.isSome()) {
-      return Value.of(inherited.value, source);
-    }
-
-    return inherited;
+  public inherited<N extends Name>(name: N): Value<Style.Inherited<N>> {
+    return this.parent.computed(name);
   }
 
   public toJSON(): Style.JSON {